#132 Upload size needs to be long

+++ src/FileSystem/DotSync.FileSystem.Application/Orchestration/IFilePusher.cs
@@ -12,5 +12,5 @@ public interface IFilePusher
     Task<IEnumerable<DotFile>> PushFilesInDir(StorageLocationType sourceType, FileSystemPath sourceRootPath,
         FileSystemPath sourcePushPath, StorageLocationType destinationType, FileSystemPath destinationPath);
 
-    Task<IEnumerable<DotFile>> PushFilesInStorage(Guid storageLocationId, string prefixFilter, int uploadLimitMb);
+    Task<IEnumerable<DotFile>> PushFilesInStorage(Guid storageLocationId, string prefixFilter, long uploadLimitMb);
 }
